+++ src/Phan/Analysis/PreOrderAnalysisVisitor.php
@@ -3,6 +3,7 @@
 namespace Phan\Analysis;
 
 use AssertionError;
+use ast;
 use ast\Node;
 use Phan\AST\ContextNode;
 use Phan\AST\UnionTypeVisitor;
@@ -22,10 +23,8 @@ use Phan\Language\FQSEN\FullyQualifiedClassName;
 use Phan\Language\FQSEN\FullyQualifiedFunctionName;
 use Phan\Language\Scope\ClosureScope;
 use Phan\Language\Type;
-use Phan\Language\Type\GenericArrayType;
 use Phan\Language\Type\VoidType;
 use Phan\Language\UnionType;
-use Phan\Library\StringUtil;
 
 /**
  * PreOrderAnalysisVisitor is where we do the pre-order part of the analysis
@@ -60,7 +59,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
     }
 
     /**
-     * Visit a node with kind `\ast\AST_CLASS`
+     * Visit a node with kind `ast\AST_CLASS`
      *
      * @param Node $node
      * A node to parse
@@ -77,7 +76,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
      */
     public function visitClass(Node $node) : Context
     {
-        if ($node->flags & \ast\flags\CLASS_ANONYMOUS) {
+        if ($node->flags & ast\flags\CLASS_ANONYMOUS) {
             $class_name =
                 (new ContextNode(
                     $this->code_base,
@@ -125,7 +124,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
     }
 
     /**
-     * Visit a node with kind `\ast\AST_METHOD`
+     * Visit a node with kind `ast\AST_METHOD`
      *
      * @param Node $node
      * A node to parse
@@ -183,7 +182,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
         }
 
         // Add $this to the scope of non-static methods
-        if (!($node->flags & \ast\flags\MODIFIER_STATIC)) {
+        if (!($node->flags & ast\flags\MODIFIER_STATIC)) {
             if (!($clazz->getInternalScope()->hasVariableWithName('this'))) {
                 throw new AssertionError("Classes must have a \$this variable.");
             }
@@ -216,7 +215,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
     }
 
     /**
-     * Visit a node with kind `\ast\AST_FUNC_DECL`
+     * Visit a node with kind `ast\AST_FUNC_DECL`
      *
      * @param Node $node
      * A node to parse
@@ -347,7 +346,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
         Func $func
     ) {
         // skip adding $this to internal scope if the closure is a static one
-        if ($func->getFlags() == \ast\flags\MODIFIER_STATIC) {
+        if ($func->getFlags() == ast\flags\MODIFIER_STATIC) {
             return;
         }
 
@@ -377,7 +376,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
 
 
     /**
-     * Visit a node with kind `\ast\AST_CLOSURE`
+     * Visit a node with kind `ast\AST_CLOSURE`
      *
      * @param Node $node
      * A node to parse
@@ -404,10 +403,10 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
         // Make the closure reachable by FQSEN from anywhere
         $code_base->addFunction($func);
 
-        if (($node->children['uses']->kind ?? null) == \ast\AST_CLOSURE_USES) {
+        if (($node->children['uses']->kind ?? null) == ast\AST_CLOSURE_USES) {
             $uses = $node->children['uses'];
             foreach ($uses->children as $use) {
-                if (!($use instanceof Node) || $use->kind != \ast\AST_CLOSURE_VAR) {
+                if (!($use instanceof Node) || $use->kind != ast\AST_CLOSURE_VAR) {
                     $this->emitIssue(
                         Issue::VariableUseClause,
                         $node->lineno ?? 0
@@ -433,7 +432,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
                 )) {
                     // If this is not pass-by-reference variable we
                     // have a problem
-                    if (!($use->flags & \ast\flags\PARAM_REF)) {
+                    if (!($use->flags & ast\flags\PARAM_REF)) {
                         Issue::maybeEmitWithParameters(
                             $this->code_base,
                             clone($context)->withLineNumberStart($use->lineno),
@@ -464,7 +463,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
                     // If this isn't a pass-by-reference variable, we
                     // clone the variable so state within this scope
                     // doesn't update the outer scope
-                    if (!($use->flags & \ast\flags\PARAM_REF)) {
+                    if (!($use->flags & ast\flags\PARAM_REF)) {
                         $variable = clone($variable);
                     }
                 }
@@ -556,10 +555,10 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
      */
     public function visitAssign(Node $node) : Context
     {
-        // In php 7.0, a **valid** parsed AST would be an \ast\AST_LIST.
-        // However, --force-polyfill-parser will emit \ast\AST_ARRAY.
+        // In php 7.0, a **valid** parsed AST would be an ast\AST_LIST.
+        // However, --force-polyfill-parser will emit ast\AST_ARRAY.
         $var_node = $node->children['var'];
-        if (Config::get_closest_target_php_version_id() < 70100 && $var_node instanceof Node && $var_node->kind === \ast\AST_ARRAY) {
+        if (Config::get_closest_target_php_version_id() < 70100 && $var_node instanceof Node && $var_node->kind === ast\AST_ARRAY) {
             $this->analyzeArrayAssignBackwardsCompatibility($var_node);
         }
         return $this->context;
@@ -595,43 +594,23 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
         if (!($value_node instanceof Node)) {
             return $context;
         }
-        if ($value_node->kind == \ast\AST_ARRAY) {
+        if ($value_node->kind == ast\AST_ARRAY) {
             if (Config::get_closest_target_php_version_id() < 70100) {
                 $this->analyzeArrayAssignBackwardsCompatibility($value_node);
             }
-            $this->inferTypesForForeachArrayDestructuring($expression_union_type->genericArrayElementTypes(), $value_node);
-
-        // Otherwise, read the value as regular variable and
-        // add it to the scope
-        } else {
-            // Filter out the non-generic types of the
-            // expression
-            $non_generic_expression_union_type =
-                $expression_union_type->iterableValueUnionType($code_base);
-
-            // Create a variable for the value
-            $variable = Variable::fromNodeInContext(
-                $value_node,
-                $context,
-                $code_base,
-                false
-            );
-
-            // If we were able to figure out the type and it's
-            // a generic type, then set its element types as
-            // the type of the variable
-            if (!$non_generic_expression_union_type->isEmpty()) {
-                $variable->setUnionType($non_generic_expression_union_type);
-            }
-
-            // Add the variable to the scope
-            $context->addScopeVariable($variable);
         }
+
+        $context = (new AssignmentVisitor(
+            $code_base,
+            $context,
+            $value_node,
+            $expression_union_type->iterableValueUnionType($code_base)
+        ))->__invoke($value_node);
 
         // If there's a key, make a variable out of that too
         $key_node = $node->children['key'];
         if ($key_node instanceof Node) {
-            if ($key_node->kind == \ast\AST_LIST) {
+            if ($key_node->kind == ast\AST_LIST) {
                 throw new NodeException(
                     $node,
                     "Can't use list() as a key element - aborting"
@@ -722,7 +701,7 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
 
     private function analyzeArrayAssignBackwardsCompatibility(Node $node)
     {
-        if ($node->flags !== \ast\flags\ARRAY_SYNTAX_LIST) {
+        if ($node->flags !== ast\flags\ARRAY_SYNTAX_LIST) {
             $this->emitIssue(
                 Issue::CompatibleShortArrayAssignPHP70,
                 $node->lineno ?? 0
@@ -738,219 +717,6 @@ class PreOrderAnalysisVisitor extends ScopeVisitor
             }
         }
     }
-
-    /**
-     * @return void
-     */
-    private function inferTypesForForeachArrayDestructuring(
-        UnionType $element_union_type,
-        Node $value_node
-    ) {
-        if ($element_union_type->hasTopLevelArrayShapeTypeInstances()) {
-            $this->inferTypesForShapedForeachArrayDestructuring($element_union_type, $value_node);
-        } else {
-            $this->inferTypesForGenericForeachArrayDestructuring($element_union_type, $value_node);
-        }
-    }
-
-    /**
-     * @param UnionType $element_union_type (The T[]|S[] types from the expression of the foreach)
-     * @param Node $value_node a node of type ast\AST_ARRAY that is in the foreach's `as` value
-     * @return void
-     */
-    private function inferTypesForShapedForeachArrayDestructuring(
-        UnionType $element_union_type,
-        Node $value_node
-    ) {
-        $key_set = [];
-
-        $expect_int_keys_lineno = false;
-        $expect_string_keys_lineno = false;
-
-        $fallback_second_order_element_type = null;
-        $get_fallback_second_order_element_type = function () use (&$fallback_second_order_element_type, $element_union_type) : UnionType {
-            return $fallback_second_order_element_type ?? ($fallback_second_order_element_type = $element_union_type->genericArrayElementTypes());
-        };
-        foreach ($value_node->children as $child_node) {
-            // $key_node = $child_node->children['key'] ?? null;
-            $value_elem_node = $child_node->children['value'] ?? null;
-
-            // for syntax like: foreach ([] as list(, $a));
-            if (!($value_elem_node instanceof Node)) {
-                continue;
-            }
-
-            // Get the key and value nodes for each
-            // array element we're assigning to
-            // TODO: Check key types are valid?
-            $key_node = $child_node->children['key'];
-
-            if ($key_node === null) {
-                $key_set[] = true;
-                \end($key_set);
-                $key_value = \key($key_set);
-
-                $expect_int_keys_lineno = $child_node->lineno;  // list($x, $y) = ... is equivalent to list(0 => $x, 1 => $y) = ...
-            } else {
-                if ($key_node instanceof Node) {
-                    $key_value = (new ContextNode($this->code_base, $this->context, $key_node))->getEquivalentPHPScalarValue();
-                } else {
-                    $key_value = $key_node;
-                }
-                if (\is_scalar($key_value)) {
-                    $key_set[$key_value] = true;
-                    if (\is_int($key_value)) {
-                        $expect_int_keys_lineno = $child_node->lineno;
-                    } elseif (\is_string($key_value)) {
-                        $expect_string_keys_lineno = $child_node->lineno;
-                    }
-                } else {
-                    $key_type = UnionTypeVisitor::unionTypeFromNode($this->code_base, $this->context, $key_node);
-                    $key_type_enum = GenericArrayType::keyTypeFromUnionTypeValues($key_type);
-                    // TODO: Warn about types that can't cast to int|string
-                    if ($key_type_enum === GenericArrayType::KEY_INT) {
-                        $expect_int_keys_lineno = $child_node->lineno;
-                    } elseif ($key_type_enum === GenericArrayType::KEY_STRING) {
-                        $expect_string_keys_lineno = $child_node->lineno;
-                    }
-                }
-            }
-
-            $variable = Variable::fromNodeInContext(
-                $value_elem_node,
-                $this->context,
-                $this->code_base,
-                false
-            );
-
-            if (\is_scalar($key_value)) {
-                $second_order_non_generic_expression_union_type = UnionTypeVisitor::resolveArrayShapeElementTypesForOffset($element_union_type, $key_value);
-                if ($second_order_non_generic_expression_union_type === null) {
-                    $second_order_non_generic_expression_union_type = $get_fallback_second_order_element_type();
-                } elseif ($second_order_non_generic_expression_union_type === false) {
-                    $this->emitIssue(
-                        Issue::TypeInvalidDimOffsetArrayDestructuring,
-                        $child_node->lineno,
-                        StringUtil::jsonEncode($key_value),
-                        (string)$element_union_type
-                    );
-                    $second_order_non_generic_expression_union_type = $get_fallback_second_order_element_type();
-                }
-            } else {
-                $second_order_non_generic_expression_union_type = $get_fallback_second_order_element_type();
-            }
-
-            // If we were able to figure out the type and it's
-            // a generic type, then set its element types as
-            // the type of the variable
-            $variable->setUnionType(
-                $second_order_non_generic_expression_union_type
-            );
-
-            $this->context->addScopeVariable($variable);
-        }
-
-        $this->checkMismatchForeachArrayDestructuringKey($expect_int_keys_lineno, $expect_string_keys_lineno, $element_union_type);
-    }
-
-    /**
-     * @param UnionType $element_union_type (The T[]|S[] types from the expression of the foreach)
-     * @param Node $value_node a node of type ast\AST_ARRAY that is in the foreach's `as` value
-     * @return void
-     */
-    private function inferTypesForGenericForeachArrayDestructuring(
-        UnionType $element_union_type,
-        Node $value_node
-    ) {
-        $scalar_array_key_cast = Config::getValue('scalar_array_key_cast');
-        $expect_string_keys_lineno = false;
-        $expect_int_keys_lineno = false;
-
-        foreach ($value_node->children as $child_node) {
-            $value_elem_node = $child_node->children['value'] ?? null;
-
-            // for syntax like: foreach ([] as list(, $a));
-            if (!($value_elem_node instanceof Node)) {
-                continue;
-            }
-
-            $key_node = $child_node->children['key'];
-            if (!$scalar_array_key_cast) {
-                if ($key_node === null) {
-                    $expect_int_keys_lineno = $child_node->lineno;  // list($x, $y) = ... is equivalent to list(0 => $x, 1 => $y) = ...
-                } else {
-                    $key_type = UnionTypeVisitor::unionTypeFromNode($this->code_base, $this->context, $key_node);
-                    $key_type_enum = GenericArrayType::keyTypeFromUnionTypeValues($key_type);
-                    // TODO: Warn about types that can't cast to int|string
-                    if ($key_type_enum === GenericArrayType::KEY_INT) {
-                        $expect_int_keys_lineno = $child_node->lineno;
-                    } elseif ($key_type_enum === GenericArrayType::KEY_STRING) {
-                        $expect_string_keys_lineno = $child_node->lineno;
-                    }
-                }
-            }
-
-            $variable = Variable::fromNodeInContext(
-                $value_elem_node,
-                $this->context,
-                $this->code_base,
-                false
-            );
-
-            // If we were able to figure out the type and it's
-            // a generic type, then set its element types as
-            // the type of the variable
-            if (!$element_union_type->isEmpty()) {
-                $second_order_non_generic_expression_union_type =
-                    $element_union_type->genericArrayElementTypes();
-
-                if (!$second_order_non_generic_expression_union_type->isEmpty()) {
-                    $variable->setUnionType(
-                        $second_order_non_generic_expression_union_type
-                    );
-                }
-            }
-
-            $this->context->addScopeVariable($variable);
-        }
-        // TODO: Issue::TypeMismatchArrayDestructuringKey,
-
-        $this->checkMismatchForeachArrayDestructuringKey($expect_int_keys_lineno, $expect_string_keys_lineno, $element_union_type);
-    }
-
-    /**
-     * TODO: Deduplicate these checks in a trait
-     * @param int|false $expect_int_keys_lineno
-     * @param int|false $expect_string_keys_lineno
-     * @param UnionType $element_union_type
-     * @return void
-     */
-    private function checkMismatchForeachArrayDestructuringKey($expect_int_keys_lineno, $expect_string_keys_lineno, UnionType $element_union_type)
-    {
-        if ($expect_int_keys_lineno !== false || $expect_string_keys_lineno !== false) {
-            $right_hand_key_type = GenericArrayType::keyTypeFromUnionTypeKeys($element_union_type);
-            if ($expect_int_keys_lineno !== false && ($right_hand_key_type & GenericArrayType::KEY_INT) === 0) {
-                Issue::maybeEmit(
-                    $this->code_base,
-                    $this->context,
-                    Issue::TypeMismatchArrayDestructuringKey,
-                    $expect_int_keys_lineno,
-                    'int',
-                    'string'
-                );
-            } elseif ($expect_string_keys_lineno !== false && ($right_hand_key_type & GenericArrayType::KEY_STRING) === 0) {
-                Issue::maybeEmit(
-                    $this->code_base,
-                    $this->context,
-                    Issue::TypeMismatchArrayDestructuringKey,
-                    $expect_string_keys_lineno,
-                    'string',
-                    'int'
-                );
-            }
-        }
-    }
-
 
     /**
      * @param Node $node